+++ example/quantity_display_mode.dart
@@ -1,15 +1,15 @@
 import 'package:human_file_size/human_file_size.dart';
+import 'package:intl/intl.dart';
 
 void main() {
+  // `SimpleQuantityDisplayMode` is the default mode.
   humanFileSize(1); // 1 B
 
-  humanFileSize(1.25); // 1.25 B
-
-  // If you were feeling extra Roman today, you could do:
+  // `IntlQuantityDisplayMode` can be used for localisation.
   humanFileSize(
-    1,
-    quantityDisplayMode: CustomQuantityDisplayMode(
-      converter: (quantity, {required unit}) => 'XIX',
+    1.234,
+    quantityDisplayMode: IntlQuantityDisplayMode(
+      numberFormat: NumberFormat.decimalPattern('pl'),
     ),
-  ); // XIX B
+  ); // 1,234 B
 }